Allow multiple task type in GitQueue
Instead of just verifying that branches exist, Git now also has the option to
test pickmes. So that we can queue this, GitQueue needs to be changed to take
a GitQueueTask which can be one of:

- VERIFY_BRANCH: check that a branch can be found and is not a duplicate
- TEST_PICKME_CONFLICT: check which (if any) branches also pickme'd for the
    same push cause merge conflicts with this branch

This also renames the update_request methods in core.git to verify_branch,
so the tests need to be updated appropriately.